+++ kura/test/org.eclipse.kura.core.test/src/test/java/org/eclipse/kura/core/internal/linux/executor/ExecutorUtilExecutionTest.java
@@ -22,14 +22,10 @@ import static org.mockito.Mockito.when;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import org.apache.commons.exec.CommandLine;
 import org.apache.commons.exec.DefaultExecutor;
 import org.apache.commons.exec.ExecuteWatchdog;
 import org.apache.commons.exec.Executor;
@@ -38,7 +34,6 @@ import org.eclipse.kura.core.linux.executor.LinuxSignal;
 import org.eclipse.kura.executor.Command;
 import org.eclipse.kura.executor.CommandStatus;
 import org.junit.Test;
-import org.mockito.ArgumentMatcher;
 
 public class ExecutorUtilExecutionTest {
 
@@ -161,33 +156,4 @@ public class ExecutorUtilExecutionTest {
         when(deMock.getWatchdog()).thenReturn(new ExecuteWatchdog(1));
     }
 
-    class CommandLineMatcher extends ArgumentMatcher<CommandLine> {
-
-        String executable;
-        List<String> arguments;
-
-        CommandLineMatcher(String executable, String[] arguments) {
-            this.executable = executable;
-            this.arguments = new ArrayList<>();
-            Arrays.asList(arguments).forEach(this.arguments::add);
-        }
-
-        @Override
-        public String toString() {
-            return executable + " " + String.join(" ", arguments);
-        }
-
-        @Override
-        public boolean matches(Object argument) {
-            boolean matched = false;
-            if (argument instanceof CommandLine) {
-                CommandLine cl = (CommandLine) argument;
-                if (this.executable.equals(cl.getExecutable()) && this.arguments.size() == cl.getArguments().length
-                        && this.arguments.containsAll(Arrays.asList(cl.getArguments()))) {
-                    matched = true;
-                }
-            }
-            return matched;
-        }
-    }
 }
