+++ kura/test/org.eclipse.kura.core.test/src/test/java/org/eclipse/kura/core/internal/linux/executor/ExecutorUtilRunningTest.java
@@ -22,19 +22,14 @@ import static org.mockito.Mockito.when;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-import org.apache.commons.exec.CommandLine;
 import org.apache.commons.exec.DefaultExecutor;
 import org.apache.commons.exec.Executor;
 import org.eclipse.kura.core.linux.executor.LinuxPid;
 import org.eclipse.kura.executor.Pid;
 import org.junit.Test;
-import org.mockito.ArgumentMatcher;
 
 public class ExecutorUtilRunningTest {
 
@@ -172,33 +167,4 @@ public class ExecutorUtilRunningTest {
         }
     }
 
-    class CommandLineMatcher extends ArgumentMatcher<CommandLine> {
-
-        String executable;
-        List<String> arguments;
-
-        CommandLineMatcher(String executable, String[] arguments) {
-            this.executable = executable;
-            this.arguments = new ArrayList<>();
-            Arrays.asList(arguments).forEach(this.arguments::add);
-        }
-
-        @Override
-        public String toString() {
-            return executable + " " + String.join(" ", arguments);
-        }
-
-        @Override
-        public boolean matches(Object argument) {
-            boolean matched = false;
-            if (argument instanceof CommandLine) {
-                CommandLine cl = (CommandLine) argument;
-                if (this.executable.equals(cl.getExecutable()) && this.arguments.size() == cl.getArguments().length
-                        && this.arguments.containsAll(Arrays.asList(cl.getArguments()))) {
-                    matched = true;
-                }
-            }
-            return matched;
-        }
-    }
 }
